refactor(negotiation-ecs): extract engine into its own Go module
Make the reuse boundary compiler-enforced. The engine core moves out of the
backend-go module into a standalone, sibling module:

- engine/ is now module github.com/erancihan/negotiation-ecs/engine (go.mod with
  zero dependencies — stdlib only). `go list -m all` there returns only itself.
- All imports rewritten from .../backend-go/engine[/ecs] to .../engine[/ecs]
  across the app, examples, and tests (git tracks the engine files as renames).
- backend-go consumes the engine via require + replace => ../engine; the Go SDK
  resolves clean (module-graph pruning keeps engine out of its graph).
- arch_test now forbids importing the entire backend-go module: the engine
  reaching into the app would require adding it to go.mod and would create a
  module cycle, so the boundary fails at build time, not just at test time.
- Makefile: test-go now also runs the engine module's tests (test-engine).
- README structure refreshed (was still referencing the deleted Ark packages);
  ROADMAP notes the extraction under a "Beyond the roadmap" section.

Both modules build and test green; headless server and all three SDKs unaffected.
