+++ playground/negotiation-ecs/backend-go/cmd/server/main.go
@@ -28,8 +28,8 @@ import (
 
 	"google.golang.org/grpc"
 
-	"github.com/erancihan/negotiation-ecs/backend-go/engine"
-	"github.com/erancihan/negotiation-ecs/backend-go/engine/ecs"
+	"github.com/erancihan/negotiation-ecs/engine"
+	"github.com/erancihan/negotiation-ecs/engine/ecs"
 	pb "github.com/erancihan/negotiation-ecs/backend-go/gen/proto/negotiationpb"
 	"github.com/erancihan/negotiation-ecs/backend-go/internal/economy"
 	"github.com/erancihan/negotiation-ecs/backend-go/internal/sim"
