+++ playground/negotiation-ecs/backend-go/internal/economy/economy.go
@@ -7,8 +7,8 @@
 package economy
 
 import (
-	"github.com/erancihan/negotiation-ecs/backend-go/engine"
-	"github.com/erancihan/negotiation-ecs/backend-go/engine/ecs"
+	"github.com/erancihan/negotiation-ecs/engine"
+	"github.com/erancihan/negotiation-ecs/engine/ecs"
 	"github.com/erancihan/negotiation-ecs/backend-go/internal/sim"
 )
