+++ playground/negotiation-ecs/backend-go/internal/integration/integration_test.go
@@ -12,8 +12,8 @@ import (
 	"sort"
 	"testing"
 
-	"github.com/erancihan/negotiation-ecs/backend-go/engine"
-	"github.com/erancihan/negotiation-ecs/backend-go/engine/ecs"
+	"github.com/erancihan/negotiation-ecs/engine"
+	"github.com/erancihan/negotiation-ecs/engine/ecs"
 	"github.com/erancihan/negotiation-ecs/backend-go/internal/economy"
 	"github.com/erancihan/negotiation-ecs/backend-go/internal/sim"
 )
