+++ playground/negotiation-ecs/backend-go/internal/sim/brains.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"math/rand"
 
-	"github.com/erancihan/negotiation-ecs/backend-go/engine/ecs"
+	"github.com/erancihan/negotiation-ecs/engine/ecs"
 )
 
 // Brains maps entities to their negotiation brains. Stored as a resource and
