+++ playground/negotiation-ecs/backend-go/internal/sim/external.go
@@ -5,7 +5,7 @@ import (
 	"encoding/json"
 	"sync"
 
-	"github.com/erancihan/negotiation-ecs/backend-go/engine/ecs"
+	"github.com/erancihan/negotiation-ecs/engine/ecs"
 )
 
 // ExternalAgents is the registry connecting remote agents to bodies they control.
