+++ playground/negotiation-ecs/backend-go/internal/sim/external_test.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"testing"
 
-	"github.com/erancihan/negotiation-ecs/backend-go/engine/ecs"
+	"github.com/erancihan/negotiation-ecs/engine/ecs"
 )
 
 func TestExternalAgentsBindIsStableAndPooled(t *testing.T) {
