+++ playground/negotiation-ecs/backend-go/internal/sim/movement.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"math"
 
-	"github.com/erancihan/negotiation-ecs/backend-go/engine"
-	"github.com/erancihan/negotiation-ecs/backend-go/engine/ecs"
+	"github.com/erancihan/negotiation-ecs/engine"
+	"github.com/erancihan/negotiation-ecs/engine/ecs"
 )
 
 // Movement is split into three systems run in order each tick:
