+++ playground/negotiation-ecs/backend-go/internal/sim/movement_test.go
@@ -6,8 +6,8 @@ import (
 	"math/rand"
 	"testing"
 
-	"github.com/erancihan/negotiation-ecs/backend-go/engine"
-	"github.com/erancihan/negotiation-ecs/backend-go/engine/ecs"
+	"github.com/erancihan/negotiation-ecs/engine"
+	"github.com/erancihan/negotiation-ecs/engine/ecs"
 )
 
 func TestSeekBrainAimsAtTarget(t *testing.T) {
