+++ playground/negotiation-ecs/backend-go/internal/sim/systems.go
@@ -3,8 +3,8 @@ package sim
 import (
 	"math"
 
-	"github.com/erancihan/negotiation-ecs/backend-go/engine"
-	"github.com/erancihan/negotiation-ecs/backend-go/engine/ecs"
+	"github.com/erancihan/negotiation-ecs/engine"
+	"github.com/erancihan/negotiation-ecs/engine/ecs"
 )
 
 // tickOf reads the current tick from the engine clock.
