+++ playground/negotiation-ecs/backend-go/internal/transport/server.go
@@ -5,8 +5,8 @@ import (
 	"log"
 	"time"
 
-	"github.com/erancihan/negotiation-ecs/backend-go/engine"
-	"github.com/erancihan/negotiation-ecs/backend-go/engine/ecs"
+	"github.com/erancihan/negotiation-ecs/engine"
+	"github.com/erancihan/negotiation-ecs/engine/ecs"
 	pb "github.com/erancihan/negotiation-ecs/backend-go/gen/proto/negotiationpb"
 	"github.com/erancihan/negotiation-ecs/backend-go/internal/sim"
 )
