+++ playground/negotiation-ecs/engine/ecs/entity.go
