+++ playground/negotiation-ecs/engine/go.mod
@@ -1,0 +1,7 @@
+// Module engine is a standalone, domain-agnostic game-engine core (ECS,
+// fixed-timestep loop, schedule, plugins, events, commands, actors, observers).
+// It depends only on the Go standard library — no third-party packages, and
+// nothing from any application that builds on it.
+module github.com/erancihan/negotiation-ecs/engine
+
+go 1.24.0
